pyproject.toml
[project]
name = "migracao-fire2pg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.96.0",
    "fdb>=2.0.2",
    "google-adk>=1.31.1",
    "litellm>=1.83.0",
    "prompt-toolkit>=3.0.52",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "sqlalchemy>=2.0.49",
]
